# Optimize .hdr loading and RGB9E5 conversion