@fire 2024-08-08
Can you post your samples as png / openexr converted from the your test cases?