@BlueCube3310 2024-08-08
Here are the samples in .exr format: [hdrsamples.zip](https://github.com/user-attachments/files/16552130/hdrsamples.zip)

I'll upload png comparisons later.