@BlueCube3310 2024-08-09
| master | PR | diff |
|--------|----|-----|
| ![falls_unopt](https://github.com/user-attachments/assets/dd4ce98a-74d2-470a-b45e-ef0e1b82e1f4) | ![falls_opt](https://github.com/user-attachments/assets/10e0ff7d-e1a1-4eba-9fce-ff8426468bf1) | ![falls_diff](https://github.com/user-attachments/assets/e6eb0c28-d093-44b5-a101-0219440a450e) |
| ![square_unpot](https://github.com/user-attachments/assets/3bc02cda-3eb6-41e8-8db9-352bcf7a2b6c) | ![square_opt](https://github.com/user-attachments/assets/79950145-6e5d-497a-aa73-663e35d74867) | ![square_diff](https://github.com/user-attachments/assets/0d1d28d4-eb01-41d8-b01a-6f2d0f41cd9f) |
| ![table-unopt](https://github.com/user-attachments/assets/01dba44f-b606-499a-97a0-0af64c6d7762) | ![table-opti](https://github.com/user-attachments/assets/7229f8c0-940d-47ea-9ae4-8cc8221285da) | ![table-diff](https://github.com/user-attachments/assets/568de96d-e1a0-4347-ab9f-fcd269083e21) |

Image sources:
https://polyhaven.com/a/crystal_falls - Greg Zaal - CC0
https://polyhaven.com/a/st_peters_square_night - Andreas Mischok - CC0
https://polyhaven.com/a/table_mountain_1 - Greg Zaal - CC0