@clayjohn 2024-08-09
> > The code looks great and the performance improvements are very exciting. I assume most of the performance benefit comes from not reading 8 bits at a time?
> 
> Interestingly, the RGBE conversion makes the biggest difference. The fast low-range optimization had the biggest impact on load speeds.

Wow, that is really surprising! But I guess it is more than just a few extra instructions, the old path moves data around a lot more too.